* notice.idをlongに変更 * PDFを日付のフォルダに格納するように変更 * PDFファイルのリンクをクリックしてプレビューできるように変更

+++ src/main/java/com/example/nbs/web/Global.java
@@ -3,6 +3,6 @@ package com.example.nbs.web;
 public class Global {
 
 
-    public static int notice_id;
+    public static long notice_id;
 
 }
